requirements.txt
fastapi
uvicorn
jinja2
python-multipart
langchain
langchain-openai
python-dotenv
markdown
